refactor: move bootstrap to its own file
