+++ cognee/tests/performance/locustio_performance_analysis.py
@@ -313,32 +313,6 @@ if __name__ == "__main__":
     import urllib.error
     import urllib.request
 
-    # Importing locust at module top monkey-patches socket/time via gevent, which
-    # makes asyncio + async sqlite hot-spin at 100% CPU. Run bootstrap in a fresh
-    # interpreter that never imports locust.
-    BOOTSTRAP_SCRIPT = """
-import asyncio
-import sys
-
-import cognee
-from cognee.modules.engine.operations.setup import setup
-from cognee.modules.users.api_key.create_api_key import create_api_key
-from cognee.modules.users.methods import create_default_user
-
-
-async def main(out_path: str) -> None:
-    await cognee.prune.prune_data()
-    await cognee.prune.prune_system(metadata=True)
-    await setup()
-    user = await create_default_user()
-    api_key_obj = await create_api_key(user, name='locust-loadtest')
-    with open(out_path, 'w') as f:
-        f.write(api_key_obj.api_key)
-
-
-asyncio.run(main(sys.argv[1]))
-"""
-
     def wait_for_server(url: str, timeout: float = 240.0) -> None:
         deadline = time.time() + timeout
         while time.time() < deadline:
@@ -354,7 +328,7 @@ asyncio.run(main(sys.argv[1]))
     key_path = tempfile.NamedTemporaryFile(suffix=".key", delete=False).name
     try:
         subprocess.run(
-            [sys.executable, "-c", BOOTSTRAP_SCRIPT, key_path],
+            [sys.executable, "-m", "bootstrap_script", key_path],
             check=True,
         )
         api_key = Path(key_path).read_text().strip()
